refactor: class locations and exceptions
