+++ rest/src/main/kotlin/org/flaxo/rest/manager/gitplag/GitplagManager.kt
@@ -7,9 +7,9 @@ import io.gitplag.gitplagapi.model.enums.GitProperty
 import io.gitplag.gitplagapi.model.input.RepositoryInput
 import okhttp3.ResponseBody
 import org.flaxo.common.Language
+import org.flaxo.model.ModelException
 import org.flaxo.model.data.Course
-import org.flaxo.moss.GitplagClient
-import org.flaxo.moss.toGitplagLanguage
+import org.flaxo.rest.manager.UnsupportedLanguageException
 import org.flaxo.rest.manager.ValidationManager
 import retrofit2.Call
 
@@ -21,16 +21,20 @@ class GitplagManager(
 ) : ValidationManager {
 
     override fun activate(course: Course) {
+        val language = Language.from(course.settings.language)
+                ?: throw UnsupportedLanguageException("Course ${course.name} does not have language property")
+        val userGithubId = course.user.githubId
+                ?: throw ModelException("Github id for ${course.user.name} user was not found")
         gitplagClient.addRepository(RepositoryInput(
                 id = -1,
                 git = GitProperty.GITHUB,
-                language = toGitplagLanguage(Language.from(course.settings.language)),
-                name = course.user.githubId!! + "/" + course.name,
+                language = toGitplagLanguage(language),
+                name = userGithubId + "/" + course.name,
                 analyzer = AnalyzerProperty.MOSS,
-                filePatterns = listOf(""".+\.java"""),
+                filePatterns = language.extensions.map { """.+\.$it""" },
                 analysisMode = AnalysisMode.FULL
         )).callUnit()
-        gitplagClient.updateRepository("github", course.user.githubId!!, course.name).callUnit()
+        gitplagClient.updateRepository("github", userGithubId, course.name).callUnit()
     }
 
     override fun deactivate(course: Course) {
